Feat: 타입 분리 및 api 호출 에러처리
