+++ src/apis/docs.ts
@@ -1,7 +1,6 @@
-import { ISearchResult } from '@/types/request';
+import { CreateDocs, ISearchResult } from '@/types/request';
 import axios from 'axios';
 import { Server } from './settings';
-import { CreateDocs } from './types';
 
 const baseURL = 'http://llwiki.p-e.kr:8000';
 
@@ -11,37 +10,46 @@ export const getSearchResult = async (keyword: string) => {
 };
 
 export const newDocs = async (body: CreateDocs) => {
-	const token = localStorage.getItem('access');
-	const result = await axios.post(`${baseURL}/docs/`, body, {
-		headers: {
-			Authorization: `Bearer ${token}`,
-		},
-	});
-	return result.data;
+	try {
+		const token = localStorage.getItem('access');
+		const result = await axios.post(`${baseURL}/docs/`, body, {
+			headers: {
+				Authorization: `Bearer ${token}`,
+			},
+		});
+		return result.data;
+	} catch (error) {
+		console.error('문서 생성 불가', error);
+		throw error;
+	}
 };
 
 export const uploadImageToServer = async (blobUrl: string) => {
-	// blob URL로부터 Blob 객체 가져오기
-	const response = await fetch(blobUrl);
-	const blob = await response.blob();
+	try {
+		// blob URL로부터 Blob 객체 가져오기
+		const response = await fetch(blobUrl);
+		const blob = await response.blob();
 
-	// Blob 객체를 File 객체로 변환
-	const file = new File([blob], 'image.png', { type: 'image/png' });
+		// Blob 객체를 File 객체로 변환
+		const file = new File([blob], 'image.png', { type: 'image/png' });
 
-	// 이미지 파일을 서버에 전송
-	let formData = new FormData();
-	formData.append('image', file);
-	const config = {
-		headers: {
-			'Content-Type': 'multipart/form-data',
-		},
-	};
-	const result = await axios.post(`${baseURL}/docs/image/`, formData, config);
-	return result.data.image;
+		// 이미지 파일을 서버에 전송
+		let formData = new FormData();
+		formData.append('image', file);
+		const config = {
+			headers: {
+				'Content-Type': 'multipart/form-data',
+			},
+		};
+		const result = await axios.post(`${baseURL}/docs/image/`, formData, config);
+		return result.data.image;
+	} catch (error) {
+		console.error('이미지 업로드 불가', error);
+		throw error;
+	}
 };
 
 export const getRecentDocs = async () => {
-	const result = await axios.get(`${baseURL}/history/recent/`);
-	console.log(result.data.data);
+	const result = await Server.get(`history/recent/`);
 	return result.data.data;
 };
